Add SSH command line parsing for terminal title
In case when remote system doesn't try to set terminal title it sticks
with old text which is misleading. Now command `ssh` (only short name)
detected and parsed to extract host name and possible login name
passed via '-l' parameter. All parameters that receive values are
expected to pass values as a separate parameter. Command split is done
according to quoting rules. The list of parameters with values are
taken from SSH manual and is subject to update according to newer
version of SSH.